% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ks.best_signiture_de.R
\name{ks.best_signiture_de}
\alias{ks.best_signiture_de}
\title{ks.best_signiture_de}
\usage{
ks.best_signiture_de(selected_miRNAs, use_mix = F)
}
\arguments{
\item{selected_miRNAs}{Vector of selected miRNAs to be checked.}

\item{use_mix}{By default (i.e. FALSE) we check the differential expression only on training dataset. If you want to check it on whole dataset (training, testing and validation dataset combined) set it to TRUE.}
}
\value{
Results of differential expression.
}
\description{
As a part of checkpoint, you may want to check the differential expression of selected miRNAs. This function uses `ks.miRNA_differential_expression()` to check the miRNAs on training dataset.
}
