% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ks.combat.R
\name{ks.combat}
\alias{ks.combat}
\title{ks.combat}
\usage{
ks.combat(danex, metadane = metadane, model = c("~ Batch", "~ Batch + Class"))
}
\arguments{
\item{danex}{Matrix with miRNA normalized expression values (e.g. TPM, deltaCt) with miRNAs in columns and cases in rows.}

\item{metadane}{Metadata with `Class` and `Batch` variable.}

\item{model}{Model of correction (application of covariates).}
}
\value{
Batch-corrected dataset.
}
\description{
Use combat to fight batch effect.
}
