% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ks.correct_miRNA_names.R
\name{ks.correct_miRNA_names}
\alias{ks.correct_miRNA_names}
\title{ks.correct_miRNA_names}
\usage{
ks.correct_miRNA_names(temp, species = "hsa")
}
\arguments{
\item{temp}{Dataset with miRNA names in columns.}
}
\value{
Corrected dataset.
}
\description{
Sometinmes, when using the dataset mapped to previous versions of miRbase we may get false mismatches due to changes in terminology.
This function uses latest version of miRbase to correct all old miRNA names to new one.
}
