% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ks.correlation_plot.R
\name{ks.correlation_plot}
\alias{ks.correlation_plot}
\title{ks.correlation_plot}
\usage{
ks.correlation_plot(
  var1,
  var2,
  labvar1,
  labvar2,
  title,
  yx = T,
  metoda = "pearson",
  gdzie_legenda = "topleft"
)
}
\arguments{
\item{var1}{First numeric variable.}

\item{var2}{Second numeric variable.}

\item{labvar1}{Label to be written on plot for fist variable.}

\item{labvar2}{Label to be written on plot for second variable.}

\item{yx}{Logical varable. TRUE if you want f(x)=x axis for reference (usefull for calibration plots).}

\item{gdzie_legenda}{Where the legend should be placed?}
}
\description{
Draw a correlation plot with reference f(x)=x axis and correlation metrics.
}
