% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ks.create_miRNA_formula.R
\name{ks.create_miRNA_formula}
\alias{ks.create_miRNA_formula}
\title{ks.create_miRNA_formula}
\usage{
ks.create_miRNA_formula(wybrane_miRy)
}
\arguments{
\item{wybrane_miRy}{Selected miRNAs as characted vector.}
}
\value{
Formula "Class ~ ..." to be used in another functions.
}
\description{
Helper function to create formula based on selected miRNAs.
}
