% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ks.deep_learning.R
\name{ks.deep_learning}
\alias{ks.deep_learning}
\title{ks.deep_learning}
\usage{
ks.deep_learning(
  selected_miRNAs = ".",
  wd = getwd(),
  SMOTE = F,
  keras_batch_size = 64,
  clean_temp_files = T,
  save_threshold_trainacc = 0.85,
  save_threshold_testacc = 0.8,
  keras_epochae = 5000,
  keras_epoch = 2000,
  keras_patience = 50,
  hyperparameters = expand.grid(layer1 = seq(3, 11, by = 2), layer2 = c(0, seq(3, 11,
    by = 2)), layer3 = c(0, seq(3, 11, by = 2)), activation_function_layer1 = c("relu",
    "sigmoid"), activation_function_layer2 = c("relu", "sigmoid"),
    activation_function_layer3 = c("relu", "sigmoid"), dropout_layer1 = c(0, 0.1),
    dropout_layer2 = c(0, 0.1), dropout_layer3 = c(0), layer1_regularizer = c(T, F),
    layer2_regularizer = c(T, F), layer3_regularizer = c(T, F), optimizer = c("adam",
    "rmsprop", "sgd"), autoencoder = c(0, 7, -7),      balanced = SMOTE, formula =
    as.character(ks.create_miRNA_formula(selected_miRNAs))[3], scaled = c(T, F),
    stringsAsFactors = F),
  keras_threads = ceiling(parallel::detectCores()/2),
  start = 1,
  end = nrow(hyperparameters),
  output_file = "deeplearning_results.csv",
  save_all_vars = F
)
}
\description{
Finding the best deep feed-forward neural network.
To be described fully...
}
