% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ks.docker.update_progress.R
\name{ks.docker.update_progress}
\alias{ks.docker.update_progress}
\title{ks.docker.update_progress}
\usage{
ks.docker.update_progress(verbose = F)
}
\description{
Helper for docker. Updates progress of analysis by 1 step.
}
