% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ks.download_tissue_miRNA_data_from_TCGA.R
\name{ks.download_tissue_miRNA_data_from_TCGA}
\alias{ks.download_tissue_miRNA_data_from_TCGA}
\title{ks.download_tissue_miRNA_data_from_TCGA}
\usage{
ks.download_tissue_miRNA_data_from_TCGA(data_folder = getwd())
}
\arguments{
\item{data_folder}{Path. Where to dowload the data?}
}
\description{
Function allowing to dowload all miRNA-seq data from TCGA.
}
