% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ks.get_benchmark.R
\name{ks.get_benchmark}
\alias{ks.get_benchmark}
\title{ks.get_benchmark}
\usage{
ks.get_benchmark(benchmark_csv = "benchmark1578929876.21765.csv")
}
\arguments{
\item{benchmark_csv}{Path to csv file.}
}
\value{
Benchmark data frame.
}
\description{
Load benchmark results from csv file.
}
