% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ks.get_miRNAs_from_benchmark.R
\name{ks.get_miRNAs_from_benchmark}
\alias{ks.get_miRNAs_from_benchmark}
\title{ks.get_miRNAs_from_benchmark}
\usage{
ks.get_miRNAs_from_benchmark(
  benchmark_csv = "benchmark1578990441.6531.csv",
  method = "fcsig"
)
}
\arguments{
\item{benchmark_csv}{Path to benchmark csv.}

\item{method}{Method of interest.}
}
\value{
Vector of miRNAs.
}
\description{
Get which miRNAs are used by which method in the benchmark.
}
