% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ks.heatmap.R
\name{ks.heatmap}
\alias{ks.heatmap}
\title{ks.heatmap}
\usage{
ks.heatmap(
  x = trainx[, 1:10],
  rlab = data.frame(Batch = dane$Batch, Class = dane$Class),
  zscore = F,
  margins = c(10, 10)
)
}
\arguments{
\item{x}{Matrix of log-transformed TPM-normalized counts with miRNAs in columns and cases in rows.}

\item{rlab}{Data frame of factors to be marked on heatmap (like batch or class). Maximum of 2 levels for every variable is supported.}

\item{zscore}{Whether to z-score values before clustering and plotting.}
}
\value{
Heatmap.
}
\description{
Draw a heatmap of selected miRNAs.
}
