% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ks.keras_create_model.R
\name{ks.keras_create_model}
\alias{ks.keras_create_model}
\title{ks.keras_create_model}
\usage{
ks.keras_create_model(
  i,
  hyperparameters,
  how_many_features = ncol(x_train_scale)
)
}
\description{
Helper function for `ks.deep_learning()`
}
