% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ks.load_extension.R
\name{ks.load_extension}
\alias{ks.load_extension}
\title{ks.load_extension}
\usage{
ks.load_extension(name = "deeplearning")
}
\arguments{
\item{name}{Name of extension to be loaded.}
}
\description{
Function for loading extension.
}
