% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ks.log.R
\name{ks.log}
\alias{ks.log}
\title{ks.log}
\usage{
ks.log(message_to_log = "Working...", logfile = "temp.log")
}
\arguments{
\item{logfile}{Path to logfile.}

\item{message}{Message to log.}
}
\description{
Logging to sink.
}
