% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ks.miRNA_differential_expression.R
\name{ks.miRNA_differential_expression}
\alias{ks.miRNA_differential_expression}
\title{ks.miRNA_differential_expression}
\usage{
ks.miRNA_differential_expression(ttpm_pofiltrze, klasy, mode = "logtpm")
}
\arguments{
\item{ttpm_pofiltrze}{matrix of log-transformed TPM-normalized miRNAs counts with miRNAs placed in `ttpm` design (i.e. columns and cases placed as rows)}

\item{klasy}{vector describing label for each case. It should contain only "Cancer" and "Control" labeles!!!!}

\item{mode}{use 'logtpm' for log(TPM) data or 'deltact' for qPCR deltaCt values. This parameters sets how the fold-change is calculated.}
}
\value{
Data frame with results.
}
\description{
The variable performes standard differential expression analysis using unpaired t-test with BH and Bonferonni correction.
It requires `ttpm_pofiltrze` object, which is a matrix of log-transformed TPM-normalized miRNAs counts with miRNAs placed in `ttpm` design (i.e. columns and cases placed as rows).
Classess should be passed as `klasy` and this should be a vector of length equal to number of rows in `ttpm_polfiltrze` and contain only "Cancer" or "Control" labels!!
The function returns the miRNAs sorted by BH-corrected p-value.
}
