% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ks.PCA.R
\name{ks.PCA}
\alias{ks.PCA}
\title{ks.PCA}
\usage{
ks.PCA(ttpm_pofiltrze, meta)
}
\arguments{
\item{ttpm_pofiltrze}{Normalizaed counts in `ttpm` format.}

\item{meta}{Factor of cases labels that should be visualized on biplot.}
}
\value{
Biplot.
}
\description{
Conduct PCA and create biplot.
}
