% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ks.PCA_3D.R
\name{ks.PCA_3D}
\alias{ks.PCA_3D}
\title{ks.PCA_3D}
\usage{
ks.PCA_3D(ttpm_pofiltrze, meta)
}
\arguments{
\item{ttpm_pofiltrze}{Normalizaed counts in `ttpm` format.}

\item{meta}{Factor of cases labels that should be visualized on biplot.}
}
\value{
Plotly 3D object.
}
\description{
Conduct PCA and create 3D scatterplot form first 3 components.
}
