% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ks.prepare_split.R
\name{ks.prepare_split}
\alias{ks.prepare_split}
\title{ks.prepare_split}
\usage{
ks.prepare_split(metadane = metadane, ttpm = ttpm_pofiltrze, train_proc = 0.6)
}
\arguments{
\item{metadane}{Metadata of cases. Must contain `Class` variable with `Cancer` and `Control` values.}

\item{ttpm}{Normalized counts used (primary data for the rest of the analysis).}

\item{train_proc}{What perc. should be kept in training dataset?}
}
\value{
The mixed dataset is return. In working directory mixed_train.csv, mixed_test.csv and mixed_valid.csv are saved. This is a crucial step in data preprocessing.
}
\description{
Create split required for `ks.miRNAselector()` and all the following functions. It is obligatory to use it.
The function devides the dataset into training, testing and validation set. Be default (as `train_proc=0.6`) 60 perc. of cases will be assigned to trainining datset.
The rest is devided into testing and validation dataset in half, ending in 60 perc. of cases in training dataset, 20 perc. of cases in testing dataset and 20 perc. of cases in validation dataset.
Metadata have to have `Class` variable, with `Cancer` and `Control` values.
}
