% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ks.process_tissue_miRNA_TCGA.R
\name{ks.process_tissue_miRNA_TCGA}
\alias{ks.process_tissue_miRNA_TCGA}
\title{ks.process_tissue_miRNA_TCGA}
\usage{
ks.process_tissue_miRNA_TCGA(
  data_folder = getwd(),
  remove_miRNAs_with_null_var = T
)
}
\arguments{
\item{data_folder}{Directory where TCGA data were downloaded.}

\item{remove_miRNAs_with_null_var}{Wheter to remove the miRNAs without any expression? Default: True}
}
\description{
Process the data downloaded from TCGA.
}
