% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ks.profileplot.R
\name{ks.profileplot}
\alias{ks.profileplot}
\title{ks.profileplot}
\usage{
ks.profileplot(
  form,
  Method.id,
  standardize = TRUE,
  interval = 10,
  by.pattern = TRUE,
  original.names = TRUE
)
}
\description{
Modified function of profileR::profileplot to correct "Method".
}
