% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ks.table.R
\name{ks.table}
\alias{ks.table}
\title{ks.table}
\usage{
ks.table(table, height = "400px")
}
\arguments{
\item{table}{Table to draw.}

\item{hight}{High (default: 400px)}
}
\description{
A wrapper to draw pretty tables in rmarkdown document
}
