% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ks.vulcano_plot.R
\name{ks.vulcano_plot}
\alias{ks.vulcano_plot}
\title{ks.vulcano_plot}
\usage{
ks.vulcano_plot(
  selected_miRNAs,
  DE = ks.miRNA_differential_expression(),
  only_label = NULL
)
}
\arguments{
\item{selected_miRNAs}{Vector of selected miRNAs.}

\item{DE}{Differential expression to get the values from. It can be generated by `ks.miRNA_differential_expression()` or `ks.best_signature_de()`}

\item{only_label}{If NULL - all selected miRNAs will be labeled. If you want to label only selected miRNAs provide a character vector with their names.}
}
\value{
Vulcano plot.
}
\description{
Draw vulcano plot of selected miRNAs.
}
