% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ks.setup.R
\name{ks.setup}
\alias{ks.setup}
\title{ks.setup}
\usage{
ks.setup(keras = TRUE, msg = TRUE)
}
\arguments{
\item{keras}{Install keras using install_keras()?}

\item{msg}{Display message if completed successfully?}
}
\description{
Run this function to be sure that everything is installed properly for miRNAselector.
}
